#236 refactor : util 성격과 상이한 도메인 service로 도메인 변경
