+++ src/test/java/com/floney/floney/book/CategoryServiceTest.java
@@ -6,7 +6,7 @@ import com.floney.floney.book.entity.RootCategory;
 import com.floney.floney.book.entity.category.BookCategory;
 import com.floney.floney.book.repository.BookRepository;
 import com.floney.floney.book.repository.category.CategoryRepository;
-import com.floney.floney.book.service.CategoryServiceImpl;
+import com.floney.floney.book.service.category.CategoryServiceImpl;
 import com.floney.floney.common.constant.Status;
 import com.floney.floney.common.exception.book.NotFoundBookException;
 import com.floney.floney.common.exception.book.NotFoundCategoryException;
